model/requirements.txt
pypdf
sentence-transformers
faiss-cpu
google-generativeai
numpy
tqdm
flask
flask-cors
python-dotenv